--mode CLI option added to change the colors.
It takes a word that can either be 'dark' or 'light'. The first one is
for terminals with a dark background while the second one is for light
backgrounds. The default mode assumes no background color and should be
fine on both.

This is part of the v0.1.7 release and fixes #3